requirements.txt
Django==2.0.6
django-extensions==2.0.7
djangorestframework==3.8.2 
djangorestframework-jwt==1.11.0
